refactor(tag settings): Rename file PE-458
