+++ src/ardrive.test.ts
@@ -13,7 +13,7 @@ import { ByteCount, UnixTime, stubTransactionID, W, FeeMultiple } from './types'
 import { readJWKFile } from './utils/common';
 import { expectAsyncErrorThrow } from '../tests/test_helpers';
 import { WalletDAO } from './wallet_dao';
-import { ArFSTagSettings } from './arfs/arfs_tag_builder';
+import { ArFSTagSettings } from './arfs/arfs_tag_settings';
 import { ArFSCostEstimator } from './pricing/arfs_cost_estimator';
 
 describe('ArDrive class', () => {
